Reorganizes the metrics code to support metrics submodules the same way we do in other sections.
