+++ endpoints/cookie_sync_test.go
@@ -15,7 +15,6 @@ import (
 	"github.com/prebid/prebid-server/config"
 	"github.com/prebid/prebid-server/gdpr"
 	"github.com/prebid/prebid-server/openrtb_ext"
-	"github.com/prebid/prebid-server/pbsmetrics"
 	"github.com/prebid/prebid-server/usersync"
 	"github.com/prebid/prebid-server/usersync/usersyncers"
 )
@@ -103,7 +102,7 @@ func doPost(body string, existingSyncs map[string]string, gdprHostConsent bool, 
 }
 
 func testableEndpoint(perms gdpr.Permissions) httprouter.Handle {
-	return NewCookieSyncEndpoint(syncersForTest(), &config.Cookie{}, perms, &pbsmetrics.DummyMetricsEngine{}, analyticsConf.NewPBSAnalytics(&config.Analytics{}))
+	return NewCookieSyncEndpoint(syncersForTest(), &config.Cookie{}, perms, &config.DummyMetricsEngine{}, analyticsConf.NewPBSAnalytics(&config.Analytics{}))
 }
 
 func syncersForTest() map[openrtb_ext.BidderName]usersync.Usersyncer {
